Fixes to compile on 64 bit Windows. Added new types H3DUInt32, H3DInt64 and H3DUInt64. Also added defines H3D_WINDOWS, H3D_WIN32 and H3D_WIN64 to indicate what platform is being compiled on.
git-svn-id: https://www.h3d.org:8090/H3DUtil/trunk@136 875a7764-6f23-0410-b93c-dbbdadb1ea73

+++ H3DUtil/src/Threads.cpp
@@ -525,7 +525,7 @@ void HapticThreadBase::synchronousHapticCB(
     }
   }
   sg_lock.lock(); 
-  haptic_threads_left = threads.size();
+  haptic_threads_left = (int) threads.size();
   if( haptic_threads_left > 0 ) {
     for( vector< HapticThreadBase *>::iterator i = threads.begin();
          i != threads.end(); i++ ) {
